refactor: 예약일 선택 본문 UI 분리
