+++ src/components/datepicker/index.tsx
@@ -1,50 +1,17 @@
 import React from 'react';
 import Header from './Header';
 import Weekdays from './Weekdays';
-import {
-  eachWeekOfInterval,
-  endOfMonth,
-  startOfMonth,
-  endOfWeek,
-  eachDayOfInterval,
-  isBefore,
-  isAfter,
-} from 'date-fns';
+import Body from './Body';
 import styled from 'styled-components';
 
 function Datepicker() {
   const today = new Date();
 
-  const startDatesOfWeeks = eachWeekOfInterval({
-    start: startOfMonth(today),
-    end: endOfMonth(today),
-  });
-
-  const isOutOfRange = (date: Date) =>
-    isBefore(date, startOfMonth(today)) || isAfter(date, endOfMonth(today));
-
   return (
     <Container>
       <Header today={today} />
       <Weekdays />
-      <Days>
-        {startDatesOfWeeks.map((startDateOfWeek) => {
-          const endDateOfWeek = endOfWeek(startDateOfWeek);
-          const days = eachDayOfInterval({
-            start: startDateOfWeek,
-            end: endDateOfWeek,
-          });
-          const daysOfWeek = days.map((day) => {
-            return isOutOfRange(day) ? (
-              <Day key={day.toDateString()}></Day>
-            ) : (
-              <Day key={day.toDateString()}>{day.getDate()}</Day>
-            );
-          });
-
-          return <DaysOfTheWeek key={startDateOfWeek.toDateString()}>{daysOfWeek}</DaysOfTheWeek>;
-        })}
-      </Days>
+      <Body today={today} />
     </Container>
   );
 }
@@ -55,19 +22,4 @@ const Container = styled.section`
   max-width: 340px;
   margin: 0 auto;
   padding: 24px;
-`;
-
-const Days = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: 24px;
-`;
-const DaysOfTheWeek = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-`;
-const Day = styled.div`
-  width: 32px;
-  text-align: center;
 `;
